+++ CHANGELOG.md
@@ -7,11 +7,13 @@ ChangeLog
 
 - **New:** Added `sortBy` method.
 - **New:** Added `switchableLabel` column option.
-- **Update** Added support for `class` attribute in toolbar buttons.
+- **New:** Added support for `class` attribute in toolbar buttons.
+- **Update:** Removed title from columns button.
 
 #### Extensions
 
 - **Update(addrbar):** Fixed clear search bug when clicking clearSearch button.
+- **Update(filter-control):** Fixed pagination server side not working bug.
 
 ### 1.21.4